Refatoração de código nas camadas

+++ backend/ecommerce_sport/src/main/java/io/github/guilhermemelo01/ecommerce_sport/repository/CategoriaRespository.java
@@ -3,5 +3,5 @@ package io.github.guilhermemelo01.ecommerce_sport.repository;
 import io.github.guilhermemelo01.ecommerce_sport.model.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriaRespository extends JpaRepository<Categoria, Integer> {
+public interface CategoriaRespository extends JpaRepository<Categoria, Long> {
 }
